introduce a reducer file
